refactor: move getLogModeFromEnvironment to platform
