--- packages/log/ts/getLogModeFromEnvironment.browser.ts
@@ -1,6 +1,0 @@
-import { LogMode } from './types.js'
-
-// istanbul ignore next
-export function getLogModeFromEnvironment(): LogMode | undefined {
-  return undefined
-}
