+++ packages/log/ts/platform/index.browser.ts
@@ -1,4 +1,4 @@
-import type { LogEntry } from '../types.js'
+import type { LogEntry, LogMode } from '../types.js'
 
 export const toInspectLogEntry = (e: LogEntry) => e
 export function isConsoleDebugAvailable() {
@@ -9,4 +9,8 @@ declare const StandardLogColor: any
 
 export function createColorLogReporter() {
   return StandardLogColor.createColorLogReporter()
+}
+
+export function getLogModeFromEnvironment(): LogMode | undefined {
+  return undefined
 }
